Hacerle algo a las variables

+++ corte_1/taller2/problemaB.java
@@ -8,7 +8,7 @@ class Main {
 
         String[] N = sc.nextLine().split(" ");
         for (int i = 0;i<N.length;i++){
-            System.out.println(N[i]);
+            System.out.println((int) N[i]);
         }
     }
 }
